Fix flaky test_timeout
We had an interesting test failure in CI where this test failed twice in
a row: we were expecting a timeout in 0.001 seconds, but both times it
took more than one second!

How can this happen? Well, when we ask a request to timeout after N
seconds, here's what we know:

 * the request will time out, eventually
 * it won't timeout before those N seconds
 * it could timeout a long time after those N seconds, especially on a
   busy CI service

Let's take advantage of those facts by specifying a request-specific
timeout that's longer than the pool-level timeout and making sure that
the timeout was long indeed.

We still test that the request-specific timeout overrides the pool-level
timeout, but will stop failing on busy CI services.