Support async content preloading
The idea here is just to make sure the existing way works, we don't
refactor anything else.

We had to take the call to `read()` out of the constructor as we can't
have await expressions there. This solution works since
`HTTPConnectionPool.urlopen()` is the only place where responses are
instantiated. (App Engine does not count as we'll remove it later.)