Migrate TestRetry tests to async
I introduced the `_test_name` helper to fix a problem with the way
the `successful_retry` endpoint works in the test dummyserver.

Indeed, that endpoint makes sure the first query fails with a 418 code
(which will trigger a retry) and then returns a 200 code. To do that, we
use a RETRY_TEST_NAMES global with the test name as a key. Now that we
run the same test five times (sync, trio, all anyio backends), we need
to put the backend in the key, otherwise a single test would succeed and
all others would fail, since the dummyserver would answer with only 200
codes.